[ENH](frontend-core): Extract frontend-core library crate
Split the HTTP server scaffolding out of chroma-frontend into a new
frontend-core library crate so a second binary (foundation-api) can
embed the same Axum scaffolding without depending on Chroma-specific
route handlers.

Moved to frontend-core:
- ac::AdmissionControlledService (mdac circuit breaker tower service)
- auth::AuthenticateAndAuthorize trait + AuthzAction / AuthzResource
- middleware: default_json_content_type, always_json_errors
- traced_json::TracedJson extractor
- errors: generic ServerError + ErrorResponse
- config: OpenTelemetryConfig, ScorecardRule, BaseServerConfig,
  load_yaml_with_env
- tracing::init_server_otel_tracing

Kept in chroma-frontend:
- All Chroma database/collection/function route handlers (server.rs)
- FrontendServer / Frontend / FrontendConfig / FrontendServerConfig
- bin/frontend_service.rs entry point
- ValidationError (has Chroma-specific variants)

chroma-frontend re-exports the moved modules at their old paths
(crate::ac, crate::auth, crate::traced_json, crate::server_middleware)
and from public lib.rs (ScorecardRule, etc.), so handler code and
external consumers (cli, python_bindings) compile unchanged.

Verification:
- cargo check --workspace passes
- cargo check --workspace --tests passes
- frontend_service binary builds and serves /healthcheck,
  /heartbeat, /version
- Unit tests pass in both crates

Refs chroma-core/hosted-chroma#7514
