+++ rust/frontend-core/src/auth.rs
