+++ rust/frontend/Cargo.toml
@@ -46,6 +46,7 @@ chroma-system = { workspace = true }
 chroma-tracing = { workspace = true, features = ["middleware"] }
 chroma-types = { workspace = true, features = ["utoipa"] }
 chroma-sqlite = { workspace = true }
+frontend-core = { workspace = true }
 utoipa = { workspace = true }
 utoipa-axum = { version = "0.2.0", features = ["debug"] }
 utoipa-swagger-ui = { version = "9", features = ["axum"] }
