+++ rust/frontend/src/config.rs
@@ -6,18 +6,15 @@ use chroma_log::config::LogConfig;
 use chroma_segment::local_segment_manager::LocalSegmentManagerConfig;
 use chroma_sqlite::config::SqliteDBConfig;
 use chroma_sysdb::{GrpcSysDbConfig, SysDbConfig};
-use chroma_tracing::{OtelFilter, OtelFilterLevel};
 use chroma_types::{default_default_knn_index, KnnIndex};
-use figment::providers::{Env, Format, Yaml};
+use figment::providers::{Format, Yaml};
 use mdac::CircuitBreakerConfig;
 use rust_embed::Embed;
 use serde::{Deserialize, Serialize};
 
-#[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct ScorecardRule {
-    pub patterns: Vec<String>,
-    pub score: u32,
-}
+// Re-export generic scaffolding config types from frontend-core so that
+// downstream consumers of `chroma_frontend::config` keep working.
+pub use frontend_core::config::{OpenTelemetryConfig, ScorecardRule};
 
 fn default_sysdb_config() -> SysDbConfig {
     SysDbConfig::Sqlite(Default::default())
@@ -109,26 +106,6 @@ impl FrontendConfig {
     }
 }
 
-fn default_otel_service_name() -> String {
-    "chromadb".to_string()
-}
-
-fn default_otel_filters() -> Vec<OtelFilter> {
-    vec![OtelFilter {
-        crate_name: "chroma_frontend".to_string(),
-        filter_level: OtelFilterLevel::Trace,
-    }]
-}
-
-#[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct OpenTelemetryConfig {
-    pub endpoint: String,
-    #[serde(default = "default_otel_service_name")]
-    pub service_name: String,
-    #[serde(default = "default_otel_filters")]
-    pub filters: Vec<OtelFilter>,
-}
-
 fn default_persist_path() -> String {
     "./chroma".to_string()
 }
@@ -214,24 +191,7 @@ impl FrontendServerConfig {
     }
 
     pub fn load_from_path(path: &str) -> Self {
-        // SAFETY(rescrv): If we cannot read the config, we panic anyway.
-        eprintln!(
-            "==========\n{}\n==========\n",
-            std::fs::read_to_string(path).unwrap()
-        );
-        // Unfortunately, figment doesn't support environment variables with underscores. So we have to map and replace them.
-        // Excluding our own environment variables, which are prefixed with CHROMA_.
-        let mut f = figment::Figment::from(
-            Env::prefixed("CHROMA_").map(|k| k.as_str().replace("__", ".").into()),
-        );
-        if std::path::Path::new(path).exists() {
-            f = figment::Figment::from(Yaml::file(path)).merge(f);
-        }
-        let res = f.extract();
-        match res {
-            Ok(config) => config,
-            Err(e) => panic!("Error loading config: {}", e),
-        }
+        frontend_core::config::load_yaml_with_env(path)
     }
 
     pub fn single_node_default() -> Self {
